+++ mail/mail-uniffi/CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   - `markMessageAsRead` now mark message as Read
   - `LabelAs` now update only LabelType::Label
+  - `WatchHandle` now properly disconnects when it's dropped.
+  - Blocking code is now allowed in the watcher callbacks.
 
 ## [0.21.4] - 2024-11-13